pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tableGen"
version = "0.1.0"
description = "Table generation and data processing project"
readme = "docs/README.md"
requires-python = ">=3.9"
authors = [
    { name = "Author", email = "email@example.com" }
]
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D"]
ignore = ["D100", "D104"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
line-length = 88
